Task-Management-API/pyproject.toml
[project]
name = "task-management-api"
version = "0.1.0"
description = "Task Management API with FastAPI, SQLModel, and Neon PostgreSQL"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "pytest>=9.0.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
